refactor:add_friend
